perf: code splitting no ApexChart para renderização assincrona
