+++ src/pages/IndexPage.vue
@@ -1,104 +1,12 @@
 <template>
   <div class="row justify-center">
     <Header />
-    <div class="row col-md-10 col-12 justify-center">
-      <VueApexCharts
-        height="350"
-        type="donut"
-        :options="chartOptionsDonut"
-        :series="series"
-        class="q-mt-xl col-md-6 col-10"
-        ref="donut"
-      />
-    </div>
+    <Chart />
   </div>
 </template>
 
 <script setup>
-import { useExpenseStore } from '../stores/expenses'
-import { useQuasar } from 'quasar'
-import { ref } from 'vue'
-import VueApexCharts from 'vue3-apexcharts'
-import Header from '../components/Header.vue'
-import { watch } from 'vue'
-const expenses = useExpenseStore()
-const $q = useQuasar()
-watch(
-  () => $q.dark.isActive,
-  (val) => {
-    if (val) {
-      chartOptionsDonut.chart.background.value = '#121212'
-    } else {
-      chartOptionsDonut.chart.background.value = '#FFF'
-    }
-  }
-)
-const chartOptionsDonut = {
-  dataLabels: {
-    enabled: false,
-  },
-  chart: {
-    background: ref('#121212'),
-  },
-  theme: {
-    palette: 'palette7',
-  },
-  responsive: [
-    {
-      breakpoint: 600,
-      options: {
-        chart: {
-          height: 300,
-        },
-        legend: {
-          show: false,
-        },
-      },
-    },
-  ],
-  legend: {
-    show: false,
-  },
-  labels: ['Entertainment', 'Health', 'Essentials'],
-  plotOptions: {
-    pie: {
-      donut: {
-        labels: {
-          value: {
-            formatter: formatCurrency,
-          },
-          show: true,
-          total: {
-            show: true,
-            fontWeight: 'bold',
-            formatter: function (w) {
-              let val = w.globals.seriesTotals
-                .reduce((a, b) => {
-                  return a + b
-                }, 0)
-                .toLocaleString()
-                .replace(/\B(?=(\d{3})+(?!\d))/g, '.')
-
-              return `R$ ${val}`
-            },
-          },
-        },
-      },
-    },
-  },
-  yaxis: {
-    labels: {
-      formatter: formatCurrency,
-    },
-  },
-}
-const series = [
-  expenses.total('entertainment'),
-  expenses.total('health'),
-  expenses.total('essentials'),
-]
-
-function formatCurrency(val) {
-  return `R$ ${val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, '.')}`
-}
+import { defineAsyncComponent } from 'vue'
+import Header from '../components/Header.vue';
+const Chart = defineAsyncComponent(() => import('../components/Chart.vue'))
 </script>
